+++ ingest/rebuild-webex.sh
@@ -8,6 +8,11 @@ mkdir $outdir
 outfile="${outdir}/rebuild-pubs.out"
 echo "CREATING ES DOCUMENTS" > $outfile
 python ./ingest-publications.py --index ${indexname} --sparql ${ENDPOINT} --spooldir ${outdir} ${outdir}/allpubs.idx  >> $outfile 2>&1
+if ! [ -s $outdir/allpubs.idx ]
+then
+   cat $outfile | mailx -s "FAILURE - rebuild-webex.sh - no index files" fis-critical@colorado.edu
+   exit
+fi
 echo "Index counts prior to run" >> $outfile
 ./idx_get_count.sh $indexname >> $outfile
 curl -XDELETE localhost:9200/${indexname} >> $outfile
